CVSL-868 move toDateTime utils method to DateTimeclass
