+++ server/routes/creatingLicences/types/dateTime.ts
@@ -3,7 +3,7 @@ import moment from 'moment'
 import { Validate } from 'class-validator'
 import type { DateString } from './dateString'
 import ValidDateString from '../../../validators/dateStringValidator'
-import SimpleTime from './time'
+import SimpleTime, { AmPm } from './time'
 import ValidSimpleTime from '../../../validators/simpleTimeValidator'
 import DateIsBefore from '../../../validators/dateIsBefore'
 import DateIsAfterExpectedReleaseDate from '../../../validators/dateIsAfterExpectedReleaseDate'
@@ -41,6 +41,17 @@ class DateTime {
     const dateTimeString = [dateString, time.hour, time.minute, time.ampm].join(' ')
     const momentDt = moment(dateTimeString, 'YYYY MM DD hh mm a')
     return momentDt.isValid() ? momentDt.format('DD/MM/YYYY HH:mm') : undefined
+  }
+
+  static toDateTime(dt: string): DateTime | undefined {
+    const momentDt = moment(dt, 'D/MM/YYYY HHmm')
+    if (!momentDt.isValid()) {
+      return undefined
+    }
+    const dateString = momentDt.format('DD/MM/YYYY')
+    const ampm = momentDt.format('a') === 'am' ? AmPm.AM : AmPm.PM
+    const simpleTime = new SimpleTime(momentDt.format('hh'), momentDt.format('mm'), ampm)
+    return DateTime.fromDateAndTime(dateString, simpleTime)
   }
 }
